+++ lib/src/widgets/dialogs/dialogs.dart
@@ -1,2 +1,3 @@
 export 'nes_confirm_dialog.dart';
 export 'nes_dialog.dart';
+export 'nes_input_dialog.dart';