+++ example/lib/gallery/sections/loading_indicators.dart
@@ -15,7 +15,44 @@ class LoadingIndicatorsSection extends StatelessWidget {
           style: theme.textTheme.displayMedium,
         ),
         const SizedBox(height: 16),
-        const _Entry(child: NesHourglassLoadingIndicator()),
+        Wrap(
+          children: [
+            const _Entry(
+              label: 'Hourglass',
+              child: NesHourglassLoadingIndicator(),
+            ),
+            const _Entry(
+              label: 'Terminal',
+              child: NesTerminalLoadingIndicator(),
+            ),
+            _Entry(
+              label: 'Jumping Icons',
+              child: NesJumpingIconsLoadingIndicator(
+                icons: [
+                  NesIcons.sword,
+                  NesIcons.shield,
+                  NesIcons.arrow,
+                  NesIcons.helm,
+                ],
+              ),
+            ),
+            _Entry(
+              label: 'Mutating Icons',
+              child: NesMutatingIconsLoadingIndicator(
+                icons: [
+                  NesIcons.sword,
+                  NesIcons.shield,
+                  NesIcons.arrow,
+                  NesIcons.helm,
+                ],
+              ),
+            ),
+            const _Entry(
+              label: 'Pixel Row',
+              child: NesPixelRowLoadingIndicator(),
+            ),
+          ],
+        ),
       ],
     );
   }
@@ -24,24 +61,31 @@ class LoadingIndicatorsSection extends StatelessWidget {
 class _Entry extends StatelessWidget {
   const _Entry({
     required this.child,
+    required this.label,
   });
 
   final Widget child;
+  final String label;
 
   @override
   Widget build(BuildContext context) {
-    return NesContainer(
-      child: Column(
-        children: [
-          child,
-          const SizedBox(height: 8),
-          Text(
-            'NesHourglassLoadingIndicator',
-            style: Theme.of(context).textTheme.bodySmall?.copyWith(
-                  fontSize: 10,
-                ),
-          ),
-        ],
+    return Padding(
+      padding: const EdgeInsets.all(8),
+      child: NesContainer(
+        height: 140,
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            child,
+            const SizedBox(height: 8),
+            Text(
+              label,
+              style: Theme.of(context).textTheme.bodySmall?.copyWith(
+                    fontSize: 10,
+                  ),
+            ),
+          ],
+        ),
       ),
     );
   }