+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # 0.9.0
  - feat: add `NesIcons.expand`
  - fix: `NesDialog` allows a generic to be informed.
- - feat: add `NesIcons.cut`
+ - feat: add `NesIcons.cut` 
+ - feat: add `NesIcons.eraser`
 
 # 0.8.0
  - feat: add `NesIcons.leftHand`